Move the wl_buffer from xwl_window to xwl_pixmap
This is cleaner, and a wl_buffer won't be destroyed until
the associated pixmap is destroyed. That way if we
use SetWindowPixmap with a pixmap with a wl_buffer already
attached, we don't have to create a new one.

This patch also allows a future api change to be able
to pend some work to do when a wl_buffer is released.

Signed-off-by: Axel Davy <axel.davy@ens.fr>